+++ packages/storybook-addon/preset.ts
@@ -1,4 +1,4 @@
-import { withModuleFederation } from './src/utils/with-module-federation-enhanced-rsbuild';
+import { withModuleFederation } from './src/utils/with-module-federation-enhanced-rsbuild.js';
 
 import type { RsbuildConfig } from '@rsbuild/core';
 import type { moduleFederationPlugin } from '@module-federation/sdk';
@@ -18,4 +18,4 @@ export default {
     });
   },
 };
-export { PLUGIN_NAME } from './src/utils/with-module-federation-enhanced-rsbuild';
+export { PLUGIN_NAME } from './src/utils/with-module-federation-enhanced-rsbuild.js';