+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/modernjsapp
 
+## 0.1.117
+
+### Patch Changes
+
+- @module-federation/enhanced@0.21.6
+
+## 0.1.116
+
+### Patch Changes
+
+- @module-federation/enhanced@0.21.5
+
+## 0.1.115
+
+### Patch Changes
+
+- Updated dependencies [a50e068]
+  - @module-federation/enhanced@0.21.4
+
 ## 0.1.114
 
 ### Patch Changes