+++ apps/router-demo/router-remote5-2005/CHANGELOG.md
@@ -1,5 +1,29 @@
 # remote5
 
+## 1.1.24
+
+### Patch Changes
+
+- Updated dependencies [b307ae0]
+  - @module-federation/rsbuild-plugin@0.21.6
+  - @module-federation/bridge-react@0.21.6
+
+## 1.1.23
+
+### Patch Changes
+
+- @module-federation/bridge-react@0.21.5
+- @module-federation/rsbuild-plugin@0.21.5
+
+## 1.1.22
+
+### Patch Changes
+
+- Updated dependencies [444db72]
+- Updated dependencies [a50e068]
+  - @module-federation/bridge-react@0.21.4
+  - @module-federation/rsbuild-plugin@0.21.4
+
 ## 1.1.21
 
 ### Patch Changes