+++ apps/router-demo/router-remote6-2006/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [b307ae0]
+  - @module-federation/rsbuild-plugin@0.21.6
+  - @module-federation/bridge-react@0.21.6
+
+## 1.1.5
+
+### Patch Changes
+
+- @module-federation/bridge-react@0.21.5
+- @module-federation/rsbuild-plugin@0.21.5
+
+## 1.1.4
+
+### Patch Changes
+
+- 444db72: fix(bridge-react): hoist BridgeWrapper to prevent component recreation
+- Updated dependencies [444db72]
+- Updated dependencies [a50e068]
+  - @module-federation/bridge-react@0.21.4
+  - @module-federation/rsbuild-plugin@0.21.4
+
 ## 1.1.3
 
 ### Patch Changes