+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @module-federation/bridge-react
 
+## 0.21.6
+
+### Patch Changes
+
+- @module-federation/sdk@0.21.6
+- @module-federation/bridge-shared@0.21.6
+
+## 0.21.5
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+  - @module-federation/bridge-shared@0.21.5
+
+## 0.21.4
+
+### Patch Changes
+
+- 444db72: fix(bridge-react): hoist BridgeWrapper to prevent component recreation
+- Updated dependencies [a50e068]
+  - @module-federation/sdk@0.21.4
+  - @module-federation/bridge-shared@0.21.4
+
 ## 0.21.3
 
 ### Patch Changes