+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/bridge-shared
 
+## 0.21.6
+
+## 0.21.5
+
+## 0.21.4
+
 ## 0.21.3
 
 ## 0.21.2