+++ packages/bridge/vue3-bridge/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @module-federation/bridge-vue3
 
+## 0.21.6
+
+### Patch Changes
+
+- @module-federation/runtime@0.21.6
+- @module-federation/sdk@0.21.6
+- @module-federation/bridge-shared@0.21.6
+
+## 0.21.5
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+  - @module-federation/runtime@0.21.5
+  - @module-federation/bridge-shared@0.21.5
+
+## 0.21.4
+
+### Patch Changes
+
+- Updated dependencies [a50e068]
+  - @module-federation/sdk@0.21.4
+  - @module-federation/runtime@0.21.4
+  - @module-federation/bridge-shared@0.21.4
+
 ## 0.21.3
 
 ### Patch Changes