+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @module-federation/esbuild
 
+## 0.0.89
+
+### Patch Changes
+
+- @module-federation/sdk@0.21.6
+
+## 0.0.88
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+
+## 0.0.87
+
+### Patch Changes
+
+- Updated dependencies [a50e068]
+  - @module-federation/sdk@0.21.4
+
 ## 0.0.86
 
 ### Patch Changes