+++ packages/modernjs/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @module-federation/modern-js
 
+## 0.21.6
+
+### Patch Changes
+
+- Updated dependencies [b307ae0]
+  - @module-federation/rsbuild-plugin@0.21.6
+  - @module-federation/runtime@0.21.6
+  - @module-federation/enhanced@0.21.6
+  - @module-federation/sdk@0.21.6
+  - @module-federation/bridge-react@0.21.6
+  - @module-federation/cli@0.21.6
+  - @module-federation/node@2.7.25
+
+## 0.21.5
+
+### Patch Changes
+
+- Updated dependencies [35c3695]
+- Updated dependencies [94d8868]
+  - @module-federation/cli@0.21.5
+  - @module-federation/sdk@0.21.5
+  - @module-federation/enhanced@0.21.5
+  - @module-federation/bridge-react@0.21.5
+  - @module-federation/node@2.7.24
+  - @module-federation/rsbuild-plugin@0.21.5
+  - @module-federation/runtime@0.21.5
+
+## 0.21.4
+
+### Patch Changes
+
+- a50e068: refactor(manifest): collect assets from build hook
+- Updated dependencies [444db72]
+- Updated dependencies [a50e068]
+  - @module-federation/bridge-react@0.21.4
+  - @module-federation/rsbuild-plugin@0.21.4
+  - @module-federation/enhanced@0.21.4
+  - @module-federation/sdk@0.21.4
+  - @module-federation/cli@0.21.4
+  - @module-federation/node@2.7.23
+  - @module-federation/runtime@0.21.4
+
 ## 0.21.3
 
 ### Patch Changes