+++ packages/node/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @module-federation/node
 
+## 2.7.25
+
+### Patch Changes
+
+- @module-federation/runtime@0.21.6
+- @module-federation/enhanced@0.21.6
+- @module-federation/sdk@0.21.6
+
+## 2.7.24
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+  - @module-federation/enhanced@0.21.5
+  - @module-federation/runtime@0.21.5
+
+## 2.7.23
+
+### Patch Changes
+
+- Updated dependencies [a50e068]
+  - @module-federation/enhanced@0.21.4
+  - @module-federation/sdk@0.21.4
+  - @module-federation/runtime@0.21.4
+
 ## 2.7.22
 
 ### Patch Changes