+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @module-federation/rsbuild-plugin
 
+## 0.21.6
+
+### Patch Changes
+
+- b307ae0: fix(rsbuild-plugin): create different temp filepath to prevent invalid manifest.exposes
+  - @module-federation/enhanced@0.21.6
+  - @module-federation/sdk@0.21.6
+  - @module-federation/node@2.7.25
+
+## 0.21.5
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+  - @module-federation/enhanced@0.21.5
+  - @module-federation/node@2.7.24
+
+## 0.21.4
+
+### Patch Changes
+
+- a50e068: refactor(manifest): collect assets from build hook
+- Updated dependencies [a50e068]
+  - @module-federation/enhanced@0.21.4
+  - @module-federation/sdk@0.21.4
+  - @module-federation/node@2.7.23
+
 ## 0.21.3
 
 ### Patch Changes