+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @module-federation/utilities
 
+## 3.1.77
+
+### Patch Changes
+
+- @module-federation/sdk@0.21.6
+
+## 3.1.76
+
+### Patch Changes
+
+- Updated dependencies [94d8868]
+  - @module-federation/sdk@0.21.5
+
+## 3.1.75
+
+### Patch Changes
+
+- Updated dependencies [a50e068]
+  - @module-federation/sdk@0.21.4
+
 ## 3.1.74
 
 ### Patch Changes